Initialized from internal project to produce an OSS variant
This package comes out of code that was integrated into a shadowserver
project I developed over three years ago. It has some rough edges
because it needed to fit into a much larger system that has sizeable
edges itself. This revamping of the package is an attempt to make the
work available for OSS adoption.